+++ src/demo/colorcanvasdemo.cpp
@@ -11,7 +11,7 @@ ColorCanvasDemo::ColorCanvasDemo(QWidget* parent, MaestroController* maestro_con
 	animation->set_timing(250);
 	animation->set_reverse(true);
 
-	Section::Layer* layer = section->set_layer(Colors::MixMode::Layer);
+	Section::Layer* layer = section->set_layer(Colors::MixMode::Overlay);
 	ColorCanvas* canvas = static_cast<ColorCanvas*>(layer->section->set_canvas(CanvasType::Type::ColorCanvas));
 
 	canvas->draw_circle(ColorPresets::BLUE, 40, 40, 40, true);